refactor bc runtime layering
