+++ src/dntker/bc/error.rs
@@ -1,0 +1,17 @@
+use std::fmt;
+
+#[derive(Debug)]
+pub enum BcError {
+    NoResult,
+    /// Parse or evaluation error
+    Error(String),
+}
+
+impl fmt::Display for BcError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            BcError::NoResult => write!(f, "No result returned"),
+            BcError::Error(msg) => write!(f, "Evaluation error: {msg}"),
+        }
+    }
+}
