+++ src/dntker/mod.rs
@@ -1,13 +1,15 @@
+#![allow(clippy::module_inception)]
+
 mod bc;
+mod buffer;
+mod history;
+mod prompt;
 mod util;
 
-use atty::Stream;
-use std::io::Write;
+mod dntker;
+pub use dntker::Dntker;
+#[cfg(test)]
+pub(crate) use dntker::FilterResult;
 
-#[cfg(target_os = "windows")]
-use winconsole::console as wconsole;
-
-use std::io::{stdout, BufWriter};
-
-include!("dntker.rs");
-include!("tests.rs");
+#[cfg(test)]
+mod tests;
